Do not ignore configured max results

+++ src/olaf_config.h
@@ -94,7 +94,7 @@
 
 		//------------ Matcher configuration
 
-		int maxResults;
+		size_t maxResults;
 
 		int searchRange;
